calender-email-assistant/requirements.txt
langchain
langchain-community
langchain-google-genai
python-dotenv
pydantic